readable plots
I figured out that the reason my plots lookd so bad was the fact that Peter Heuer conflates all maxima with all minima in his CR39 library, so I was reading tracks with contrast and eccentricity *over* 35 and 15, respectively.  I switched that up and added a few more touches, and now the plots actually look pretty good, and the data are immediately apparent.