revise usage flow
I tried to use this for more than a few scanfiles at a time, and with multiple fingers on each CPS, and it turns out it was nigh unusable!  so I revised it to more effectively handle large numbers of files (and to use argparser instead of the weird thing I made up).  now you don't even need to specify the finger(s); it just figures it out on its own!